temp renaming settingsContext
